Move Mobius base classes to a library module
