+++ app/src/debug/java/org/simple/clinic/di/DebugCrashReporterModule.kt
@@ -1,8 +1,8 @@
 package org.simple.clinic.di
 
 import org.simple.clinic.crash.CrashReporterModule
-import org.simple.clinic.crash.NoOpCrashReporter
 import org.simple.clinic.facility.FacilityRepository
+import org.simple.clinic.platform.crash.NoOpCrashReporter
 import org.simple.clinic.user.UserSession
 
 class DebugCrashReporterModule : CrashReporterModule() {
