+++ app/src/main/java/org/simple/clinic/crash/CrashReporterModule.kt
@@ -3,6 +3,7 @@ package org.simple.clinic.crash
 import dagger.Module
 import dagger.Provides
 import org.simple.clinic.facility.FacilityRepository
+import org.simple.clinic.platform.crash.CrashReporter
 import org.simple.clinic.user.UserSession
 
 @Module
