+++ app/src/main/java/org/simple/clinic/crash/SentryCrashReporter.kt
@@ -5,13 +5,15 @@ import io.reactivex.schedulers.Schedulers.io
 import io.sentry.Sentry
 import io.sentry.android.AndroidSentryClientFactory
 import io.sentry.event.BreadcrumbBuilder
-import org.simple.clinic.crash.Breadcrumb.Priority.ASSERT
-import org.simple.clinic.crash.Breadcrumb.Priority.DEBUG
-import org.simple.clinic.crash.Breadcrumb.Priority.ERROR
-import org.simple.clinic.crash.Breadcrumb.Priority.INFO
-import org.simple.clinic.crash.Breadcrumb.Priority.VERBOSE
-import org.simple.clinic.crash.Breadcrumb.Priority.WARN
 import org.simple.clinic.facility.FacilityRepository
+import org.simple.clinic.platform.crash.Breadcrumb
+import org.simple.clinic.platform.crash.Breadcrumb.Priority.ASSERT
+import org.simple.clinic.platform.crash.Breadcrumb.Priority.DEBUG
+import org.simple.clinic.platform.crash.Breadcrumb.Priority.ERROR
+import org.simple.clinic.platform.crash.Breadcrumb.Priority.INFO
+import org.simple.clinic.platform.crash.Breadcrumb.Priority.VERBOSE
+import org.simple.clinic.platform.crash.Breadcrumb.Priority.WARN
+import org.simple.clinic.platform.crash.CrashReporter
 import org.simple.clinic.user.UserSession
 import org.simple.clinic.util.filterAndUnwrapJust
 import java.util.Date
