+++ app/src/main/java/org/simple/clinic/remoteconfig/RemoteConfigSync.kt
@@ -2,7 +2,7 @@ package org.simple.clinic.remoteconfig
 
 import io.reactivex.Completable
 import io.reactivex.Single
-import org.simple.clinic.crash.CrashReporter
+import org.simple.clinic.platform.crash.CrashReporter
 import org.simple.clinic.sync.BatchSize
 import org.simple.clinic.sync.ModelSync
 import org.simple.clinic.sync.SyncConfig
