+++ app/src/main/java/org/simple/clinic/settings/changelanguage/ChangeLanguageScreen.kt
@@ -13,8 +13,8 @@ import io.reactivex.rxkotlin.ofType
 import kotlinx.android.synthetic.main.screen_change_language.view.*
 import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.activity.TheActivity
-import org.simple.clinic.crash.CrashReporter
 import org.simple.clinic.mobius.MobiusDelegate
+import org.simple.clinic.platform.crash.CrashReporter
 import org.simple.clinic.router.screen.ScreenRouter
 import org.simple.clinic.settings.Language
 import org.simple.clinic.settings.SettingsRepository
