+++ mobius-base/src/main/AndroidManifest.xml
@@ -1,2 +1,1 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="org.simple.clinic.mobius" />
+<manifest package="org.simple.clinic.mobius" />
