+++ simple-platform/build.gradle
@@ -24,10 +24,4 @@ android {
 
 dependencies {
   implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$versions.kotlin"
-  implementation project(":simple-platform")
-
-  api "com.spotify.mobius:mobius-core:$versions.mobius"
-  api "com.spotify.mobius:mobius-rx2:$versions.mobius"
-  api "com.spotify.mobius:mobius-android:$versions.mobius"
-  api "com.spotify.mobius:mobius-extras:$versions.mobius"
 }
