+++ simple-platform/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="org.simple.clinic.platform" />
